chore: Code improvements
