+++ packages/core/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@bundly/ares-core",
-  "version": "0.2.0-beta.9",
+  "version": "0.2.0-beta.10",
   "description": "Client library for interacting with the Internet Computer",
   "author": "Bundly",
   "license": "MIT",
