+++ packages/core/src/client/index.ts
@@ -1,4 +1,2 @@
 export * from "./client";
 export * from "./client.types";
-export * from "./identity-provider";
-export * from "./incomplete-ed25519-key-identity";
