--- packages/core/src/identity-providers/internet-identity/index.ts
@@ -1,2 +1,0 @@
-export * from "./internet-identity";
-export * from "./internet-identity.types";
