--- packages/core/src/identity-providers/internet-identity/internet-identity.types.ts
@@ -1,4 +1,0 @@
-export type InternetIdentityCreateOptions = {
-  providerUrl?: string;
-  maxTimeToLive?: bigint;
-};
