+++ packages/core/src/identity/index.ts
@@ -1,0 +1,4 @@
+export * from "./providers";
+export * from "./identity-manager";
+export * from "./identity-provider";
+export * from "./incomplete-ed25519-key-identity";
