+++ packages/core/src/storage/index.ts
@@ -1,3 +1,2 @@
 export * from "./local-storage";
 export * from "./storage.interface";
-export * from "./idb-storage";
