+++ packages/react-native/src/identity-providers/internet-identity-react-native/index.ts
@@ -1,2 +1,1 @@
 export * from "./internet-identity-react-native";
-export * from "./internet-identity-react-native.types";
