--- packages/react-native/src/identity-providers/internet-identity-react-native/internet-identity-react-native.types.ts
@@ -1,10 +1,0 @@
-export type InAppBrowser = {
-  open: (url: string) => void;
-  close: () => void;
-};
-
-export type InternetIdentityReactNativeConfig = {
-  providerUrl: string;
-  appLink: string;
-  inAppBrowser: InAppBrowser;
-};
