+++ packages/react/package.json
@@ -31,7 +31,7 @@
   },
   "dependencies": {
     "@bundly/ares-rest": "^0.1.1-beta.0",
-    "@bundly/ares-core": "^0.2.0-beta.9",
+    "@bundly/ares-core": "^0.2.0-beta.10",
     "@dfinity/agent": "^1.3.0",
     "@dfinity/auth-client": "^1.3.0",
     "@dfinity/identity": "^1.3.0",
